model fitting compressed into one main file; removed individual fitting files
